fix: support metadata in deferred calls execution (#1724)
* fix: metadata support in RestDeferredCall execution

* codegen for scala & java

* samples and docs

* fixing scala test

* fixing spring test

* testing java test with mocks

* fixing java tests

* removing var from the codegen method

* fixing scala codegen

* fixing java codegen

* fixing IT test

* missing space in the comment

* missing codegen updates

+++ samples/java-protobuf-valueentity-shopping-cart/src/main/java/com/example/shoppingcart/domain/ShoppingCart.java
@@ -117,7 +117,12 @@ public class ShoppingCart extends AbstractShoppingCart {
   @Override
   public Effect<Empty> removeCart(
           ShoppingCartDomain.Cart currentState, ShoppingCartApi.RemoveShoppingCart removeShoppingCart) {
-    return effects().deleteEntity().thenReply(Empty.getDefaultInstance());
+    var userRole = commandContext().metadata().get("Role").get();
+    if (userRole.equals("Admin")) {
+      return effects().deleteEntity().thenReply(Empty.getDefaultInstance());
+    } else {
+      return effects().error("Only admin can remove the cart");
+    }
   }
